Add !null checks to Bitmap.recycle() calls
previously was not checking whether a View was null before recycling it - resulting in a crash if you entered in a URL that passed the response code and formatting checks, and had images to download, but did not return a valid bitmap.

An example of this was an Imgur gallery, that has multiple images and formatting on it. The URLwould pass the checks but return a null ImageView.

I still need to figure out a way to either capture multiple images in that situation, or find where I can return an error in a textview. Currently the !null checks ensure that the app doesn't crash in that instance, however it doesn't return any useful information to the user.

Also some misc fixes and code clean up.