Implement ImageDL object into AsyncTask
Changed the Async Task to return a new ImageDL object. The ImageDL object contains URI's to the thumbnail and fullsize downloaded image locations, as well as a created Image or Text view. This implementation also includes changes to ImageSiphon to download thumbs and create an image view based off of the thumb saved to storage. Implementing the ImageDL object will allow for future features like recalling current ImageViews onPause() and sending URIs in bundles. We will also be able to view the fullsize images within the application in the future.