Support onSave and onRestoreInstanceState
Added downloaded thumbnail URIs to an arraylist to save in bundles when app is paused, and now restoring images onRestore. Now the app no longer loses its image views/images when rotating or switching out of the app. I also increased functionality of the Clear button to also remove the downloaded images from storage and to wipe the thumbnail URI arraylist clean and display a message to the user.