Fix Progress Bar
Added counters to initial for loop (iterating over entered links, and calling an AsyncTask for each link) and then a counter for completed images onPostExecute(). Comparing these values and setting the visibility of the progress bar to GONE if they match, allows the progress bar to stay up the full duration of downloading all the images, instead of just the initial one.