Add Functionality to Down File Button
Adds a file reader to the app as well as a links.txt file in the assets folder. When pressing the Down File button the app will read the file and parse the strings into a string array - in an async task, and returns them to the main thread to be downloaded. This makes demoing the app easier, instead of having to copy/paste links everytime.